Fifth Commit

+++ src/ViewStaff.js
@@ -5,12 +5,12 @@ export default function ViewStaff() {
   const [state, setState] = useState([]);
   const [data, setData] = useState();
   function dataRemove(id){
-  
+    if(window.confirm("Are you sure want to delete ?")){
       axios.delete(`http://localhost:3000/user/${id}`).then((response)=>{
       }).catch((error)=>
       console.log(error))
       window.location.reload()
-    
+    }
     }
   useEffect(() => {
     axios.get("http://localhost:3000/user").then((response) => {
@@ -22,14 +22,15 @@ export default function ViewStaff() {
   useEffect(() => {
     setData(
       state.map((data) => {
-        return (
+        return data.role==="Librian"?(
           <tr>
             <td>{data.name}</td>
             <td>{data.email}</td>
             <td>{data.address}</td>
-            <td><i class="fa-solid fa-trash" onClick={() =>dataRemove(data.id)}></i></td>
+            <td>{data.phone}</td>
+            <td><i class="fa-solid fa-circle-xmark" onClick={() =>dataRemove(data.id)} style={{color:"red",fontSize:20}}></i></td>
           </tr>
-        );
+        ):"";
       })
     );
   }, [state]);
@@ -37,18 +38,19 @@ export default function ViewStaff() {
   return (
     
     <div>
-      <div className="reg-form">
+      <div className="table">
         <h2>
           <center>
             <b>View Staff Here!</b>
           </center>
-        </h2>
-        <table border={2} align="center">
+        </h2><br></br>
+        <table className="viewtable" border={2} align="center">
           <thead>
             <tr>
             <th>Name</th>
             <th>Email</th>
             <th>Address</th>
+            <th>Mobile No.</th>
             <th>Action</th>
             </tr>
           </thead>
